Use safer left-shift of negative integers
Left-shift of a negative integer using a signed integer type is
undefined behaviour in C although GCC and its derivates allow it.
Casting to an unsigned integer type preserves the two's complement
representation and allows for well defined left-shift behaviour.

+++ src/iso8825_strings.c
@@ -90,7 +90,7 @@ static int iso8825_asn1_integer_get_string(
 		x = 0;
 	}
 	for (unsigned int i = 0; i < tlv->length; ++i) {
-		x = (x << 8) | tlv->value[i];
+		x = ((uint64_t)x << 8) | tlv->value[i];
 	}
 
 	snprintf(str, str_len, "%"PRId64, x);
